+++ tests/frontend/org/qcri/affinityplanner/TestReconfigurationPredictor.java
@@ -12,24 +12,24 @@ import edu.brown.utils.ProjectType;
  */
 public class TestReconfigurationPredictor extends BaseTestCase {
 
-    private double[] load_predictions_arr_1 = new double[]{ 
+    private long[] load_predictions_arr_1 = new long[]{
             100, 200, 300, 400, 500, 600, 600, 500, 400, 300, 200, 100, 
             100, 200, 300, 400, 500, 600, 600, 500, 400, 300, 200, 100 };
-    private double capacity_per_node_1 = 100;
+    private long capacity_per_node_1 = 100;
     private int nodes_start_1 = 1;
     private int db_migration_time_1 = 2;
 
-    private double[] load_predictions_arr_2 = new double[]{ 
+    private long[] load_predictions_arr_2 = new long[]{
             600, 500, 400, 300, 200, 100, 100, 200, 300, 400, 500, 600, 
             600, 500, 400, 300, 200, 100, 100, 200, 300, 400, 500, 600 };
-    private double capacity_per_node_2 = 100;
+    private long capacity_per_node_2 = 100;
     private int nodes_start_2 = 6;
     private int db_migration_time_2 = 2;
     
-    private double[] load_predictions_arr_3 = new double[]{ 
+    private long[] load_predictions_arr_3 = new long[]{
             100, 100, 100, 100, 100, 600, 100, 100, 100, 100, 100, 100, 
             100, 100, 100, 100, 100, 400, 100, 100, 100, 100, 100, 100 };
-    private double capacity_per_node_3 = 100;
+    private long capacity_per_node_3 = 100;
     private int nodes_start_3 = 1;
     private int db_migration_time_3 = 2;
 
@@ -38,8 +38,8 @@ public class TestReconfigurationPredictor extends BaseTestCase {
         super.setUp(ProjectType.B2W);
     }
 
-    private void checkCorrect(ReconfigurationPredictor predictor, ArrayList<Move> moves, 
-            double[] load_predictions_arr, boolean print) {
+    private void checkCorrect(ReconfigurationPredictor predictor, ArrayList<Move> moves,
+                              long[] load_predictions_arr, boolean print) {
         Move prev_move = null;
         if (print) System.out.println("   Machines \tCapacity \tLoad");
         for (Move move : moves) {
@@ -47,7 +47,7 @@ public class TestReconfigurationPredictor extends BaseTestCase {
             if (prev_move != null) {
                 int reconfig_time = predictor.reconfigTime(prev_move.nodes, move.nodes);
                 for (int i = 1; i < move.time - prev_move.time; ++i) {
-                    double effectiveCap = predictor.effectiveCapacity(i, reconfig_time, prev_move.nodes, move.nodes);
+                    long effectiveCap = predictor.effectiveCapacity(i, reconfig_time, prev_move.nodes, move.nodes);
                     assertTrue(effectiveCap >= load_predictions_arr[prev_move.time+i]);
                     if (print) System.out.println(prev_move.time+i + ":   \t" + effectiveCap + " \t" + load_predictions_arr[prev_move.time+i]);
                 }
@@ -59,8 +59,8 @@ public class TestReconfigurationPredictor extends BaseTestCase {
     }
 
     public void testBestMoves1() throws Exception {
-        ArrayList<Double> load_predictions = new ArrayList<>();
-        for (double prediction : load_predictions_arr_1) load_predictions.add(prediction);
+        ArrayList<Long> load_predictions = new ArrayList<>();
+        for (long prediction : load_predictions_arr_1) load_predictions.add(prediction);
 
         // defaults
         ReconfigurationPredictor predictor = new ReconfigurationPredictor(capacity_per_node_1, db_migration_time_1);
@@ -84,8 +84,8 @@ public class TestReconfigurationPredictor extends BaseTestCase {
     }
 
     public void testBestMoves2() throws Exception {
-        ArrayList<Double> load_predictions = new ArrayList<>();
-        for (double prediction : load_predictions_arr_2) load_predictions.add(prediction);
+        ArrayList<Long> load_predictions = new ArrayList<>();
+        for (long prediction : load_predictions_arr_2) load_predictions.add(prediction);
 
         // defaults
         ReconfigurationPredictor predictor = new ReconfigurationPredictor(capacity_per_node_2, db_migration_time_2);
@@ -109,8 +109,8 @@ public class TestReconfigurationPredictor extends BaseTestCase {
     }
     
     public void testBestMoves3() throws Exception {
-        ArrayList<Double> load_predictions = new ArrayList<>();
-        for (double prediction : load_predictions_arr_3) load_predictions.add(prediction);
+        ArrayList<Long> load_predictions = new ArrayList<>();
+        for (long prediction : load_predictions_arr_3) load_predictions.add(prediction);
 
         // defaults
         ReconfigurationPredictor predictor = new ReconfigurationPredictor(capacity_per_node_3, db_migration_time_3);
@@ -139,8 +139,8 @@ public class TestReconfigurationPredictor extends BaseTestCase {
     }
 
     public void testBestMoves4() throws Exception {
-        ArrayList<Double> load_predictions = new ArrayList<>();
-        double load = 100;
+        ArrayList<Long> load_predictions = new ArrayList<>();
+        long load = 100;
         // first hump
         for (int i = 0; i < 2000; ++i) {
             load_predictions.add(load);
@@ -200,7 +200,7 @@ public class TestReconfigurationPredictor extends BaseTestCase {
         ReconfigurationPredictor predictor = new ReconfigurationPredictor(8000, 500);
 
         ArrayList<Move> moves = predictor.bestMoves(load_predictions, 2);
-        double load_predictions_arr[] = new double[load_predictions.size()];
+        long load_predictions_arr[] = new long[load_predictions.size()];
         for (int i = 0; i < load_predictions_arr.length; ++i) load_predictions_arr[i] = load_predictions.get(i);
         checkCorrect(predictor, moves, load_predictions_arr, false);
